{{ cookiecutter.project_slug }}/requirements.txt
aioredis==1.3.1
amqp==2.5.2
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.11.1
Automat==0.8.0
billiard==3.6.1.0
celery==4.3.0
certifi==2018.8.24
cffi==1.13.2
channels==2.3.1
channels-redis==2.4.1
chardet==3.0.4
constantly==15.1.0
coverage==4.5.4
coveralls==1.8.2
cryptography==2.8
daphne==2.4.0
defusedxml==0.5.0
Django==2.2.4
django-allauth==0.37.1
django-cors-middleware==1.3.1
django-environ==0.4.5
django-extensions==2.2.5
django-filter==2.0.0
django-nose==1.4.6
django-oauth-toolkit==1.2.0
django-redis==4.9.0
django-rest-auth==0.9.3
django-stubs==1.0.2
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
djangorestframework-stubs==0.4.2
docopt==0.6.2
filelock==3.0.12
gunicorn==19.9.0
hiredis==1.0.1
hyperlink==19.0.0
idna==2.7
importlib-metadata==0.19
incremental==17.5.0
kombu==4.6.6
mock==3.0.5
msgpack==0.6.2
mypy==0.720
mypy-extensions==0.4.1
nose==1.3.7
oauthlib==2.1.0
packaging==19.1
Pillow==6.1.0
pluggy==0.12.0
psycopg2-binary==2.7.5
py==1.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
PyHamcrest==1.9.0
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.2
python3-openid==3.1.0
pytz==2019.2
redis==2.10.6
requests==2.19.1
requests-oauthlib==1.0.0
service-identity==18.1.0
six==1.11.0
sqlparse==0.3.0
toml==0.10.0
tox==3.13.2
Twisted==19.10.0
txaio==18.8.1
typed-ast==1.4.0
typing-extensions==3.7.4
urllib3==1.23
vine==1.3.0
virtualenv==16.7.2
zipp==0.5.2
zope.interface==4.7.1
